third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/sandbox-in-http-header.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe src="resources/sandbox.php?sandbox=allow-top-navigation"></iframe>
<script>
window.onload = () => {
    test(() => {
        assert_equals(document.getElementsByTagName("iframe")[0].contentDocument, null);
    }, "contentDocument should return null for sandboxed iframe");
};
</script>
